+++ packages/expo-cellular/android/src/main/java/expo/modules/cellular/CellularGeneration.kt
@@ -1,6 +1,8 @@
 package expo.modules.cellular
 
-enum class CellularGeneration(val value: Int) {
+import expo.modules.kotlin.types.EnumArgument
+
+enum class CellularGeneration(val value: Int) : EnumArgument {
   UNKNOWN(0),
   CG_2G(1),
   CG_3G(2),