+++ packages/expo-crypto/android/src/main/java/expo/modules/crypto/DigestAlgorithm.kt
@@ -1,6 +1,8 @@
 package expo.modules.crypto
 
-enum class DigestAlgorithm(val value: String) {
+import expo.modules.kotlin.types.EnumArgument
+
+enum class DigestAlgorithm(val value: String) : EnumArgument {
   MD5("MD5"),
   SHA1("SHA-1"),
   SHA256("SHA-256"),