+++ packages/expo-image-picker/android/src/main/java/expo/modules/imagepicker/ImagePickerOptions.kt
@@ -9,6 +9,7 @@ import expo.modules.imagepicker.contracts.CameraContractOptions
 import expo.modules.imagepicker.contracts.ImageLibraryContractOptions
 import expo.modules.kotlin.records.Field
 import expo.modules.kotlin.records.Record
+import expo.modules.kotlin.types.EnumArgument
 
 internal class ImagePickerOptions : Record, Serializable {
   @Field
@@ -41,7 +42,7 @@ internal class ImagePickerOptions : Record, Serializable {
   fun toImageLibraryContractOptions() = ImageLibraryContractOptions(this)
 }
 
-internal enum class MediaTypes(val value: String) {
+internal enum class MediaTypes(val value: String) : EnumArgument {
   IMAGES("Images"),
   VIDEOS("Videos"),
   ALL("All");