+++ packages/expo-modules-core/android/src/test/java/expo/modules/kotlin/types/EnumTypeConverterTest.kt
@@ -5,15 +5,15 @@ import com.google.common.truth.Truth
 import org.junit.Test
 
 class EnumTypeConverterTest {
-  enum class EnumWithoutParameter {
+  enum class EnumWithoutParameter : EnumArgument {
     VALUE1, VALUE2, VALUE3
   }
 
-  enum class EnumWithInt(val value: Int) {
+  enum class EnumWithInt(val value: Int) : EnumArgument {
     VALUE1(1), VALUE2(2), VALUE3(3)
   }
 
-  enum class EnumWithString(val value: String) {
+  enum class EnumWithString(val value: String) : EnumArgument {
     VALUE1("value1"), VALUE2("value2"), VALUE3("value3")
   }